# openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Task API",
      "version": "1.0.0",
      "description": "A simple task manager API"
    },
    "paths": {
      "/": {
        "get": {
          "summary": "API info",
          "responses": {
            "200": {
              "description": "Returns API name, version and endpoints"
            }
          }
        }
      },
      "/health": {
        "get": {
          "summary": "Health check",
          "responses": {
            "200": {
              "description": "Returns status ok"
            }
          }
        }
      },
      "/tasks": {
        "get": {
          "summary": "Get all tasks",
          "responses": {
            "200": {
              "description": "List of all tasks"
            }
          }
        },
        "post": {
          "summary": "Create a new task",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "Attend class"
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Task created"
            },
            "400": {
              "description": "Title is required"
            }
          }
        }
      },
      "/tasks/{id}": {
        "get": {
          "summary": "Get task by id",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Single task"
            },
            "404": {
              "description": "Task not found"
            }
          }
        },
        "put": {
          "summary": "Update a task",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": {
                      "type": "string",
                      "example": "Complete assignment"
                    },
                    "done": {
                      "type": "boolean",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Task updated"
            },
            "400": {
              "description": "Empty or invalid body"
            },
            "404": {
              "description": "Task not found"
            }
          }
        },
        "delete": {
          "summary": "Delete a task",
          "parameters": [
            {
              "in": "path",
              "name": "id",
              "required": true,
              "schema": {
                "type": "integer"
              }
            }
          ],
          "responses": {
            "204": {
              "description": "Task deleted"
            },
            "404": {
              "description": "Task not found"
            }
          }
        }
      }
    }
  }